Menambahkan hak akses role event_coordinator
